requirements.txt
aiofiles==24.1.0
aiogram==3.21.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
alembic==1.16.3
annotated-types==0.7.0
attrs==25.3.0
certifi==2025.7.9
frozenlist==1.7.0
greenlet==3.2.3
idna==3.10
magic-filter==1.0.12
Mako==1.3.10
MarkupSafe==3.0.2
multidict==6.6.3
phonenumbers==9.0.9
propcache==0.3.2
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
pytz==2025.2
SQLAlchemy==2.0.41
typing-inspection==0.4.1
typing_extensions==4.14.1
yarl==1.20.1
